+++ examples/radar/plot_turbine.py
@@ -10,7 +10,7 @@ class to create your own version.
 Each blade of the turbine represents the statistics for the skill.
 While the blade is split at the point of the individual player's skill level.
 
-If you like this idea give `Soumyajit Bose <https://twitter.com/Soumyaj15209314>`_ a follow
+If you like this idea follow `Soumyajit Bose <https://twitter.com/Soumyaj15209314>`_
 on Twitter, as I borrowed some of his ideas for this chart.
 """
 import pandas as pd
@@ -23,7 +23,8 @@ import matplotlib.pyplot as plt
 # Creating some random data
 # -------------------------
 # Here we create some random data from a truncated normal distribution
-# In real life the values would be an array or dataframe of shape number of players * number of skills
+# In real life the values would be an array or dataframe of
+# shape number of players * number of skills
 lower, upper, mu, sigma = 0, 1, 0.35, 0.25
 X = stats.truncnorm((lower - mu) / sigma, (upper - mu) / sigma, loc=mu, scale=sigma)
 # for 1000 people and 11 skills
@@ -38,7 +39,7 @@ params = ['Expected goals', 'Total shots',
 # set up a dataframe with the random values
 df = pd.DataFrame(values)
 df.columns = params
-# in real-life you'd probably have a string column for the player name
+# in real-life you'd probably have a string column for the player name,
 # but we will use numbers here
 df['player_name'] = np.arange(1000)
 
@@ -46,8 +47,8 @@ df['player_name'] = np.arange(1000)
 # Instantiate the Radar Class
 # ---------------------------
 # We will instantiate a radar object and set the lower and upper bounds.
-# For miscontrols/ dispossessed it is better to have a lower number so we
-# will flip the statistic by adding the paramater to ``lower_is_better``.
+# For miscontrols/ dispossessed it is better to have a lower number, so we
+# will flip the statistic by adding the parameter to ``lower_is_better``.
 
 # create the radar object with an upper and lower bound of the 5% and 95% quantiles
 low = df[params].quantile(0.05).values
@@ -109,6 +110,48 @@ range_labels = radar.draw_range_labels(ax=axs['radar'], fontsize=15,
                                        fontproperties=fm.prop, zorder=2)
 param_labels = radar.draw_param_labels(ax=axs['radar'], fontsize=15,
                                        fontproperties=fm.prop, zorder=2)
+
+# adding a title and endnote
+title1_text = axs['title'].text(0.01, 0.65, 'Random player', fontsize=25,
+                                fontproperties=fm.prop, ha='left', va='center')
+title2_text = axs['title'].text(0.01, 0.25, 'Team', fontsize=20,
+                                fontproperties=fm.prop,
+                                ha='left', va='center', color='#B6282F')
+title3_text = axs['title'].text(0.99, 0.65, 'Turbine Chart', fontsize=25,
+                                fontproperties=fm.prop, ha='right', va='center')
+title4_text = axs['title'].text(0.99, 0.25, 'Position', fontsize=20,
+                                fontproperties=fm.prop,
+                                ha='right', va='center', color='#B6282F')
+endnote_text = axs['endnote'].text(0.99, 0.5, 'Inspired By StatsBomb', fontsize=15,
+                                   fontproperties=fm.prop, ha='right', va='center')
+
+##############################################################################
+# Mixing with Radars
+# ------------------
+# You can also mix and match the different elements of Radars and Turbines.
+
+# creating the figure using the grid function from mplsoccer:
+fig, axs = grid(figheight=14, grid_height=0.915, title_height=0.06, endnote_height=0.025,
+                title_space=0, endnote_space=0, grid_key='radar', axis=False)
+
+# plot the turbine plot
+radar.setup_axis(ax=axs['radar'])
+# plot the turbine blades. Here we give the player_Values and
+# the value for all players shape=(1000, 11)
+turbine_output = radar.turbine(player_values, df[params].values, ax=axs['radar'],
+                               kwargs_inner={'edgecolor': '#d4d4d4', 'color': '#81b8fb'},
+                               kwargs_outer={'facecolor': '#eeeeee', 'edgecolor': '#d4d4d4'})
+# plot some dashed rings and the labels for the values and parameter names
+rings_inner = radar.draw_circles(ax=axs['radar'], facecolor='None',
+                                 edgecolor='black', linestyle='--')
+range_labels = radar.draw_range_labels(ax=axs['radar'], fontsize=15,
+                                       fontproperties=fm.prop, zorder=12)
+param_labels = radar.draw_param_labels(ax=axs['radar'], fontsize=15,
+                                       fontproperties=fm.prop, zorder=2)
+# overlay the radar
+radar_output = radar.draw_radar(player_values, ax=axs['radar'],
+                                kwargs_radar={'facecolor': '#9dc7ff', 'alpha': 0.7},
+                                kwargs_rings={'facecolor': '#bbd8ff', 'alpha': 0.7})
 
 # adding a title and endnote
 title1_text = axs['title'].text(0.01, 0.65, 'Random player', fontsize=25,